Split operators
